+++ readium/navigator/r2-navigator/src/main/java/org/readium/r2/navigator/Experimental.kt
@@ -10,3 +10,8 @@ package org.readium.r2.navigator
 @Retention(AnnotationRetention.BINARY)
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.TYPEALIAS, AnnotationTarget.PROPERTY)
 annotation class ExperimentalDecorator
+
+@RequiresOptIn(message = "The new Audiobook navigator is still experimental. The API may be changed in the future without notice.")
+@Retention(AnnotationRetention.BINARY)
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.TYPEALIAS, AnnotationTarget.PROPERTY)
+annotation class ExperimentalAudiobook